+++ tests/functional/test_inference.py
@@ -20,7 +20,6 @@ MA2 process.
 """
 
 
-
 def setup_ma2_with_informative_data():
     true_params = OrderedDict([('t1', .6), ('t2', .2)])
     n_obs = 100
@@ -95,6 +94,7 @@ def test_smc():
     assert res.populations[-1].n_batches < 6
 
 
+@pytest.mark.usefixtures('skip_travis')  # very, very slow in Travis, but ok locally
 @slow
 @pytest.mark.usefixtures('with_all_clients')
 def test_BOLFI():
@@ -105,9 +105,10 @@ def test_BOLFI():
     log_d = NodeReference(m['d'], state=dict(_operation=np.log), model=m, name='log_d')
 
     bolfi = elfi.BOLFI(log_d, initial_evidence=20, update_interval=10, batch_size=5,
-                       bounds=[(-2,2)]*len(m.parameters))
-    res = bolfi.infer(300)
-    assert bolfi.target_model.n_evidence == 300
+                       bounds=[(-2,2), (-1, 1)])
+    n = 300
+    res = bolfi.infer(n)
+    assert bolfi.target_model.n_evidence == n
     acq_x = bolfi.target_model._gp.X
 
     # check_inference_with_informative_data(res, 1, true_params, error_bound=.2)
@@ -115,24 +116,23 @@ def test_BOLFI():
     assert np.abs(res['samples']['t2'] - true_params['t2']) < 0.2
 
     # Test that you can continue the inference where we left off
-    res = bolfi.infer(310)
-    assert bolfi.target_model.n_evidence == 310
-    assert np.array_equal(bolfi.target_model._gp.X[:300,:], acq_x)
+    res = bolfi.infer(n+10)
+    assert bolfi.target_model.n_evidence == n+10
+    assert np.array_equal(bolfi.target_model._gp.X[:n,:], acq_x)
 
     post = bolfi.infer_posterior()
 
-    post_ml, _ = post.ML
-    post_map, _ = post.MAP
+    post_ml = post.ML
+    post_map = post.MAP
     vals_ml = dict(t1=np.array([post_ml[0]]), t2=np.array([post_ml[1]]))
     check_inference_with_informative_data(vals_ml, 1, true_params, error_bound=.2)
     vals_map = dict(t1=np.array([post_map[0]]), t2=np.array([post_map[1]]))
     check_inference_with_informative_data(vals_map, 1, true_params, error_bound=.2)
 
-    # Commented out because for some reason, this is very, very slow in Travis
-    # n_samples = 100
-    # n_chains = 4
-    # res_sampling = bolfi.sample(n_samples, n_chains=n_chains)
-    # check_inference_with_informative_data(res_sampling.samples, n_samples//2*n_chains, true_params, error_bound=.2)
+    n_samples = 400
+    n_chains = 4
+    res_sampling = bolfi.sample(n_samples, n_chains=n_chains)
+    check_inference_with_informative_data(res_sampling.samples, n_samples//2*n_chains, true_params, error_bound=.2)
 
     # check the cached predictions for RBF
     x = np.random.random((1, len(true_params)))
@@ -147,3 +147,9 @@ def test_BOLFI():
     grad_cached_mu, grad_cached_var = bolfi.target_model.predictive_gradients(x)
     assert(np.allclose(grad_mu, grad_cached_mu))
     assert(np.allclose(grad_var, grad_cached_var))
+
+    # test calculation of prior logpdfs
+    true_logpdf_prior = ma2.CustomPrior1.logpdf(x[0, 0], 2)
+    true_logpdf_prior += ma2.CustomPrior2.logpdf(x[0, 1], x[0, 0,], 1)
+
+    assert np.isclose(true_logpdf_prior, post._logprior_density(x[0, :]))