+++ tests/unit/test_examples.py
@@ -9,8 +9,10 @@ def test_bdm(recwarn):
     """Currently only works in unix-like systems and with a cloned repository"""
     cpp_path = ee.bdm.get_sources_path()
 
+    do_cleanup = False
     if not os.path.isfile(cpp_path + '/bdm'):
         os.system('make -C {}'.format(cpp_path))
+        do_cleanup = True
 
     assert os.path.isfile(cpp_path + '/bdm')
 
@@ -31,4 +33,6 @@ def test_bdm(recwarn):
     # TODO: test the correctness of the result
 
     os.system('rm ./bdm')
+    if do_cleanup:
+        os.system('rm {}/bdm'.format(cpp_path))